Added firebase support
Moved config.js to firestore.js, and added utility functions to that file. In each tsx, instead of using the axios() call directly, we now call the utility function relevant to that task (addPost(), getAllPosts() etc).
Any time we wish to implement another data source, we simply add a file to the /src/services directory, implement any initialization, and export the desired functionality.