Changes to teh way signInViaGoogle and signOut work, correct on Nav components as needed
so basically, rather than having the firestore be aware of the history thing, I changed the Nav.js component to expect four optional properties: two you already had (login and logout), and two I added (loginCallback and logoutCallback). The callback functions will now be run in the signInViaGoogle or signOut function's .then(...), which means we can redirect to different places, depending on the page. Further, we could abstract that a little more, and make it so that, for each page, if the user isn't logged in, we display a plug and the login Nav, and if they ARE logged in, we display the page normally with the logout Nav.
Not a significant change, but it lets us do some cool things. For example, if we needed to clean up localStorage when the user logs out, or if we wanted to set some value on the database (like a 'lastActive' timestamp on the logging out user), or redirect to some other page. This gives us flexibility.