Working redirect  back to a given page after login
Currently, when the user logs in, they're taken to a specific page. We should, with the callback system in place, be able to return to the place the user was headed to (or to the page they last had open, if that's desired). Trying to grok the logic to make that happen.