Image uploader working
Based of the FilePond image uploader, we have a react component, ImageUploader, that implements the FilePond component and allows us to upload one or more images. These are stored in the images useState hook, which we upload with our post as an images array.

That images array (at this point) is the firebase storage id of the image, but we could just as easily have changed that to image urls had we wanted. I think, by using the storage id, we can later add delete functionality.