Improved integration into the post
imageUploader now returns an object to the onRequestSave function. This is inserting an object into the array of objects. For the image, that object contains an id and url for that image.

We may want to add keywords, alt text, default image status, etc... but for tonight that works.

**ALSO**: at this point, we can start integrating those images into the PostsPage and PostDetail. We now have a `post.data.images` array, whicih we can use to retrieve the images themselves from our storage!

This is working, though it could use a little refinement. Still to be done: things like keywords/default/resize, and the ability to delete images from a post. Looking into that now...