FInished merging image-uploader branch
re-Deleted components/CreatePost/CreatePost.tsx and components/PostsPage/PostsPage.tsx (as these are now in pages). Merged the ImageUploader component into the CreatePost page, and modified the PostsPage to display images in either of two ways: if the given product has an imageUrl, use that to display the one image. If the given product has an images array, use *that* to display all images. Thus, we support the older versions as well.

further, modified getAllPosts in firestore to retrieve them in order, latest post first.