Reply functions added to firebase, integrating forms
Working on integrating a modal dialog, so when the user clicks 'Request item', they get a dialog that lets them save a message. Additionally (I think this is a firebase function), the original poster gets a text/email/notification, if we've set that up.

But at this point, working in the functionality of making a damn modal work, just in the HTML and CSS.

Haven't yet hooked it to the "replies" calls in firebase yet.