The Ant.design modal and form, working
Requesting an item is working, and only available if the user is logged in.

The 'Sign in to request' and 'Request item' buttons are part of the component, clicking on them opens the modal. I'm thinking once a user submits a request, we can probably remove that component from that post for them - they'll simply see the reply they've made.

It was a bit of a pain, and looking at the changed files, I'm thinking... huh. That was remarkably little writing for a decent-sized bit of brain matter...