Making a notes document, to try to organize my thoughts.
I want a way to update the database when the schema changes, but I only want to change the records we actually reference. That was what todays thoughts were about.
Going forward, we need to create a users collection, storing the userId, username, maybe avatar if given. This will give us something we can refer to in other documents/collections.
Also, we need to update the posts soon. These will need to include a collection of keyterms. Doing this as a collection, we can find all documents with a given keyterm in their sub-collection, making them searchable by term. I think that's our workaround for not being able to search the text (yet).